+++ src/container-registry/get-docker-credentials/get-docker-credentials.ts
@@ -2,6 +2,7 @@ import { IResponse, IContext } from '../../types';
 
 export interface IGetDockerCredentialsApiRequest {
   can_write?: boolean;
+  expiry_seconds?: number;
 }
 
 export interface IGetDockerCredentialsApiResponse {
@@ -14,9 +15,13 @@ export const getDockerCredentials = ({
   httpClient,
 }: IContext) => ({
   can_write = false,
+  expiry_seconds,
 }: IGetDockerCredentialsApiRequest): Promise<Readonly<GetDockerCredentialsResponse>> => {
   const path = '/registry/docker-credentials';
-  const query_params = {read_write: can_write};
+  const query_params = {
+    expiry_seconds,
+    read_write: can_write,
+  };
   const url = `${path}`;
 
   return httpClient.get<IGetDockerCredentialsApiResponse>(url, {params: query_params});